feat: merge options inside every month component
