+++ packages/donate-button/src/components/EveryMonth/every-month.tsx
@@ -9,6 +9,7 @@ import Company from 'src/components/Images/Company';
 import {Styled} from 'src/components/Styled';
 import DonationsContext, {AnimationValue} from 'src/contexts/donations-context';
 import OptionsContext from 'src/contexts/options-context';
+import {getFinalOptions} from 'src/helpers/final-options';
 import {
 	DefaultFrequency,
 	DonateButtonOptions,
@@ -16,10 +17,12 @@ import {
 } from 'src/helpers/options-types';
 
 interface EveryMonthProps {
-	options: DonateButtonOptions;
+	options: Partial<DonateButtonOptions>;
 	hide: () => void;
 }
 const EveryMonth = ({options, hide}: EveryMonthProps) => {
+	const finalOptions: DonateButtonOptions = getFinalOptions(options);
+
 	const hideOnWrapperClick: JSXInternal.MouseEventHandler<Element> = (
 		event
 	) => {
@@ -28,16 +31,17 @@ const EveryMonth = ({options, hide}: EveryMonthProps) => {
 		}
 	};
 
-	const isMonthlyDefault = options.defaultMode !== DefaultFrequency.ONE_TIME;
+	const isMonthlyDefault =
+		finalOptions.defaultMode !== DefaultFrequency.ONE_TIME;
 	const [monthlyDonation, setMonthlyDonation] = useState(isMonthlyDefault);
 
-	const defaultLevelIdx = options.monthly.levels.findIndex(
+	const defaultLevelIdx = finalOptions.monthly.levels.findIndex(
 		(level) => level.default
 	);
-	const defaultLevel = options.monthly.levels[defaultLevelIdx];
+	const defaultLevel = finalOptions.monthly.levels[defaultLevelIdx];
 
 	const [donationAmount, setDonationAmount] = useState(
-		defaultLevel?.amount ?? options.monthly.levels[0].amount
+		defaultLevel?.amount ?? finalOptions.monthly.levels[0].amount
 	);
 	const [customDonation, setCustomDonation] = useState('');
 	const [customInputError, setCustomInputError] = useState('');
@@ -46,7 +50,7 @@ const EveryMonth = ({options, hide}: EveryMonthProps) => {
 		defaultLevelIdx
 	]);
 	// Custom must be the last level
-	const monthlyLevels = [...options.monthly.levels].sort((a, b) =>
+	const monthlyLevels = [...finalOptions.monthly.levels].sort((a, b) =>
 		Number.isNaN(Number(b.amount)) ? -1 : 0
 	);
 
@@ -154,9 +158,9 @@ const EveryMonth = ({options, hide}: EveryMonthProps) => {
 			<div>
 				<div className="wrapper" onClick={hideOnWrapperClick}>
 					<div className="close" onClick={hideOnWrapperClick} />
-					<OptionsContext.Provider value={options}>
+					<OptionsContext.Provider value={finalOptions}>
 						<DonationsContext.Provider value={donationsContextValue}>
-							{options.mode.toUpperCase() === LayoutMode.SPLIT && (
+							{finalOptions.mode.toUpperCase() === LayoutMode.SPLIT && (
 								<div className="widget widget--split">
 									<Donations
 										monthlyDonation={monthlyDonation}
@@ -180,7 +184,7 @@ const EveryMonth = ({options, hide}: EveryMonthProps) => {
 									</div>
 								</div>
 							)}
-							{options.mode.toUpperCase() === LayoutMode.SINGLE && (
+							{finalOptions.mode.toUpperCase() === LayoutMode.SINGLE && (
 								<div className="widget widget--single">
 									<Donations
 										monthlyDonation={monthlyDonation}
